Moved util.kt to utils package
