+++ src/main/kotlin/commands/RoleInfoCommand.kt
@@ -5,6 +5,7 @@ import net.ayataka.kordis.entity.server.role.Role
 import org.kamiblue.botkt.*
 import org.kamiblue.botkt.utils.MessageSendUtils.error
 import org.kamiblue.botkt.utils.Colors
+import org.kamiblue.botkt.utils.pretty
 
 object RoleInfoCommand : Command("roleinfo") {
     init {
