+++ src/main/kotlin/utils/ReactionUtils.kt
@@ -10,7 +10,6 @@ import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.RequestBody.Companion.toRequestBody
-import org.kamiblue.botkt.getAuthToken
 import org.kamiblue.botkt.utils.StringUtils.uriEncode
 
 @Suppress("UNUSED")
