+++ src/main/kotlin/utils/util.kt
@@ -1,16 +1,14 @@
-package org.kamiblue.botkt
+package org.kamiblue.botkt.utils
 
 import com.google.gson.Gson
-import net.ayataka.kordis.entity.message.Message
 import net.ayataka.kordis.entity.server.permission.PermissionSet
-import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.OkHttpClient
 import okhttp3.Request
-import okhttp3.RequestBody.Companion.toRequestBody
+import org.kamiblue.botkt.AuthConfig
 import org.kamiblue.botkt.ConfigManager.readConfigSafe
-import org.kamiblue.botkt.utils.MessageSendUtils.error
+import org.kamiblue.botkt.ConfigType
+import org.kamiblue.botkt.CounterConfig
 import org.kamiblue.botkt.utils.StringUtils.toHumanReadable
-import org.l1ving.api.issue.Issue
 import java.util.concurrent.TimeUnit
 
 /**
